Finish PoC for InstUI team
This commit starts and finishes the PoC to show the InstUI team the
Modal a11y issue I discovered. The issue is as follows:

If you are using the InstUI Modal component and try to render an alert
to a user immediately after closing the modal, perhaps to inform them
that they're changes were successful or unsuccessful, the alert will not
be read out by screenreaders. This is likely happening due to the Modal
returning focus to the element that opened it, which effectively
"steamrolls" the alert, as the screenreader thinks that the user has
already started to tab away as soon as it starts reading the alert.